Change request logic for the smallest footprint
bacon: test
resolves: OKTA-89232
