+++ jquery/index.js
@@ -1,0 +1,2 @@
+var jqueryRequest = require('./jqueryRequest');
+module.exports = require('../lib/clientBuilder')(jqueryRequest);
