+++ package.json
@@ -19,9 +19,10 @@
     "lint": "eslint .",
     "lint:fix": "eslint --fix .",
     "lint:report": "eslint -f checkstyle -o build2/reports/lint/eslint-checkstyle-result.xml .",
-    "test": "webpack --config webpack.test.config.js && grunt test",
+    "test": "grunt test",
     "build": "webpack --config webpack.config.js",
     "build:AMDJqueryQ": "webpack --config webpack.AMDJqueryQ.config.js",
+    "build:tests": "webpack --config webpack.test.config.js",
     "package": "grunt",
     "ci-update-package": "./node_modules/@okta/ci-update-package/bin/ci-update-package.js",
     "ci-pkginfo:dataload": "./node_modules/@okta/ci-pkginfo/bin/ci-pkginfo.js -t dataload"
@@ -33,10 +34,15 @@
     "auth",
     "login"
   ],
+  "dependencies": {
+    "Base64": "0.3.0",
+    "q": "1.4.1",
+    "reqwest": "2.0.5",
+    "xhr2": "0.1.3"
+  },
   "devDependencies": {
     "@okta/ci-pkginfo": "1.2.0",
     "@okta/ci-update-package": "1.4.0",
-    "Base64": "0.3.0",
     "eslint": "2.10.2",
     "grunt": "0.4.5",
     "grunt-cli": "1.2.0",
@@ -46,8 +52,6 @@
     "lodash": "4.10.0",
     "moment": "2.12.0",
     "normalize-url": "1.4.1",
-    "q": "1.4.1",
-    "reqwest": "2.0.5",
     "semver": "5.1.0",
     "sinon": "1.16.0",
     "webpack": "1.13.0"
