+++ reqwest/index.js
@@ -1,0 +1,2 @@
+var reqwestRequest = require('./reqwestRequest');
+module.exports = require('../lib/clientBuilder')(reqwestRequest);
